+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Added
+
+- Action button allows user to export tabular results to file.
+- Tool call allows user to ask the model to export tabular results to file.
+
 ## [0.2.3] - 2025-01-24
 
 ### Added